entrega_02/voip_integration/requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
python-dotenv